refactor: split 'settings' Vuex store into 'categories' and 'views'
