+++ src/store/index.js
@@ -2,7 +2,8 @@ import Vue from 'vue';
 import Vuex from 'vuex';
 import activity from './modules/activity';
 import buckets from './modules/buckets';
-import settings from './modules/settings';
+import categories from './modules/categories';
+import views from './modules/views';
 //import createLogger from '../../../src/plugins/logger';
 
 Vue.use(Vuex);
@@ -13,7 +14,8 @@ export default new Vuex.Store({
   modules: {
     activity,
     buckets,
-    settings,
+    categories,
+    views,
   },
   strict: debug,
   //  plugins: debug ? [createLogger()] : [],
