+++ src/views/settings/CategorizationSettings.vue
@@ -37,27 +37,27 @@ export default {
     CategoryEditTree,
   },
   computed: {
-    ...mapGetters('settings', ['classes_hierarchy']),
-    ...mapState('settings', ['classes_unsaved_changes']),
+    ...mapGetters('categories', ['classes_hierarchy']),
+    ...mapState('categories', ['classes_unsaved_changes']),
   },
   mounted() {
-    this.$store.dispatch('settings/load');
+    this.$store.dispatch('categories/load');
   },
   methods: {
     addClass: function () {
-      this.$store.commit('settings/addClass', {
+      this.$store.commit('categories/addClass', {
         name: ['New class'],
         rule: { type: 'regex', regex: 'FILL ME' },
       });
     },
     saveClasses: async function () {
-      await this.$store.dispatch('settings/save');
+      await this.$store.dispatch('categories/save');
     },
     resetClasses: async function () {
-      await this.$store.dispatch('settings/load');
+      await this.$store.dispatch('categories/load');
     },
     restoreDefaultClasses: async function () {
-      await this.$store.commit('settings/restoreDefaultClasses');
+      await this.$store.commit('categories/restoreDefaultClasses');
     },
   },
 };
